requirements.txt
boto3==1.16.30
PyYAML==5.3.1
requests==2.22.0
